+++ examples/svelte/src/routes/examples/customnode/+page.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import type { ChangeEventHandler } from 'svelte/elements';
 	import { writable } from 'svelte/store';
 	import {
 		SvelteFlow,