+++ packages/react/src/components/Handle/index.tsx
@@ -1,10 +1,10 @@
 import {
-  memo,
-  forwardRef,
   type HTMLAttributes,
   type MouseEvent as ReactMouseEvent,
   type TouchEvent as ReactTouchEvent,
   type ForwardedRef,
+  memo,
+  forwardRef,
 } from 'react';
 import cc from 'classcat';
 import { shallow } from 'zustand/shallow';